### 2025-11-19 #8573 @snowformatics
> [@snowformatics](https://github.com/snowformatics) [@vbarrera](https://github.com/vbarrera) [@camillescott](https://github.com/camillescott) [@genomematt](https://github.com/genomematt) [@Ebedthan](https://github.com/Ebedthan), [@telatin](https://github.com/telatin) - Would you like to review this submission for JOSS? Thank you for considering this invitation.

@fboehm
Thank you very much for the invitation! I would be interested in reviewing the tool for JOSS. I work with such datasets and tools on a daily basis, so I can test the software thoroughly. I also have a number of relevant datasets that I can use for comparison, and I’m happy to review the text of the paper as well.

I should mention that I am not a GO programmer (I primarily work in Python), but I can provide detailed feedback on usability and overall tool quality.

If this fits the reviewer requirements for JOSS, you can count me in.
