### 2025-11-22 #8573 @fluhus
Thank you @snowformatics and @telatin for agreeing to review this. I'm looking forward to your feedback!

Please note, I pushed some changes that accumulated over the past few months since submitting this manuscript. **That includes a new release of the binaries.**

@fboehm, I also updated the performance comparison section in the manuscript. I have two questions about it:
1. I am not sure if this section aligns with JOSS's format. Should I keep it there?
   * Overall I'd like to (somewhere) explain the algorithm in detail and show comparison results. What's the usual practice for that with software submitted to JOSS? Should I create a separate manuscript for that?
2. Would it be possible to update the manuscript URL here for the reviewers? FWIW [this](https://github.com/fluhus/blini/actions/runs/19587127379/artifacts/4646419470) is the link I got from the GitHub action.

Thanks!
