### 2025-12-02 #9494 @snowformatics
I added my review in this [issue](https://github.com/fluhus/blini/issues/1).
