### 2025-12-10 #9494 @snowformatics
Dear @fboehm,

The author has addressed all of my comments and suggestions in a constructive manner, resulting in clear improvements to both the tool’s features and the manuscript. Only minor editorial issues remain, which I have noted. Subject to these small corrections, my review is complete.

I would also like to note that Blini is a useful and practical tool, and I am already using it in my own pipeline as a fast pre-screening step, which further highlights its applicability.

Please note that due to the Christmas holidays, I will be unavailable from December 17 to January 7, 2026.

Best regards
Stefanie
