### 2025-12-10 b3fd913
Add testing section to README
