### 2025-12-10 6de9ab2
Extend paper/README and minor updates to benchmark commands
