### 2025-12-10 e248aaa
Add testdata/fasta structure description
