### 2025-12-12 fa96305
Mention zenodo repo in paper README
