### 2025-12-15 b27f4de
Bump release version to 1.0.0
