### 2025-12-15 #9494 @fluhus
Thank you @snowformatics and @telatin for a fruitful review!

@fboehm, thank you for the prompt response. I have completed the tasks you mentioned. Please let me know if there is anything else you need from me.

---

  **Done**
  **v1.0.0**
  **10.5281/zenodo.17942839**
  **Done**
  **Done**

---

Thanks!
