### 2025-12-16 4a545d0
A few editorial changes to paper
