### 2025-12-17 #9494 @fboehm
@fluhus - please see the output above. You need to add dois to the references.
