### 2025-12-17 451ce0c
Add DOIs to paper bib
