+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume={17},
   pages={1--14},
   year={2016},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi={10.1186/s13059-016-0997-x}
 }
 
 @article{steinegger2018clustering,
@@ -16,7 +17,8 @@
   number={1},
   pages={2542},
   year={2018},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41467-018-04964-5}
 }
 
 @article{altschul1990basic,
@@ -27,7 +29,8 @@
   number={3},
   pages={403--410},
   year={1990},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/S0022-2836(05)80360-2}
 }
 
 @article{pruitt2007ncbi,
@@ -38,7 +41,8 @@
   number={suppl\_1},
   pages={D61--D65},
   year={2007},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkl842}
 }
 
 @article{brown2016sourmash,
@@ -48,7 +52,8 @@
   volume={1},
   number={5},
   pages={27},
-  year={2016}
+  year={2016},
+  doi={10.21105/joss.00027}
 }
 
 @article{pasolli2019extensive,
@@ -59,7 +64,8 @@
   number={3},
   pages={649--662},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cell.2019.01.001}
 }
 
 @article{simmonds2024cz,
@@ -68,5 +74,6 @@
   journal={bioRxiv},
   pages={2024--02},
   year={2024},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi={10.1101/2024.02.29.579666}
 }
