### 2025-12-17 #9494 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
