### 2025-12-22 b7fe30e
Edits to paper
