### 2025-12-22 16af2b5
Increase font size in figures
