+++ paper/plots.py
@@ -6,6 +6,7 @@ from matplotlib import pyplot as plt
 from myplot import ctx
 
 OUT_DIR = '../testdata/output'
+SIZE_RATIO = [0.5, 0.75]
 
 
 def load_table(f: str):
@@ -26,22 +27,21 @@ def load_table(f: str):
 
 def plot_search_small():
     d = load_table('results/results_search.txt')
-    print(d)
     d['name'] = [x.replace(' ', '\n') for x in d['name']]
-    with ctx(f'{OUT_DIR}/search_time', sizeratio=0.75):
+    with ctx(f'{OUT_DIR}/search_time', sizeratio=SIZE_RATIO):
         plt.bar(d['name'][:2], d['time_seconds'][:2])
         plt.bar(d['name'][2:4], d['time_seconds'][2:4])
         plt.bar(d['name'][4:], d['time_seconds'][4:])
         plt.yscale('log')
         plt.ylabel('Average time (s)')
         plt.xticks(rotation=90)
-    with ctx(f'{OUT_DIR}/search_found', sizeratio=0.75):
+    with ctx(f'{OUT_DIR}/search_found', sizeratio=SIZE_RATIO):
         plt.bar(d['name'][:2], d['source_found'][:2])
         plt.bar(d['name'][2:4], d['source_found'][2:4])
         plt.bar(d['name'][4:], d['source_found'][4:])
         plt.ylabel('Successful matches (out of 100)')
         plt.xticks(rotation=90)
-    with ctx(f'{OUT_DIR}/search_others', sizeratio=0.75):
+    with ctx(f'{OUT_DIR}/search_others', sizeratio=SIZE_RATIO):
         plt.bar(d['name'][:2], d['others_found'][:2])
         plt.bar(d['name'][2:4], d['others_found'][2:4])
         plt.bar(d['name'][4:], d['others_found'][4:])
@@ -51,7 +51,6 @@ def plot_search_small():
 
 def plot_search_big():
     d = load_table('results/results_search_big.txt')
-    print(d)
     d['name'] = [x.replace(' (', '\n(') for x in d['name']]
     with ctx('results/search_big', sizeratio=0.75):
         plt.bar(d['name'][:3], d['time_seconds'][:3])
@@ -65,19 +64,18 @@ def plot_search_big():
 
 def plot_clust(tag: str):
     d = load_table(f'results/results_clust_{tag}.txt')
-    print(d)
     d['name'] = [x.replace(' (', '\n(') for x in d['name']]
-    with ctx(f'{OUT_DIR}/clust_{tag}_time', sizeratio=0.75):
+    with ctx(f'{OUT_DIR}/clust_{tag}_time', sizeratio=SIZE_RATIO):
         plt.bar(d['name'][:4], d['time_seconds'][:4])
         plt.bar(d['name'][4:], d['time_seconds'][4:])
         plt.ylabel('Average time (s)')
         plt.xticks(rotation=90)
-    with ctx(f'{OUT_DIR}/clust_{tag}_mem', sizeratio=0.75):
+    with ctx(f'{OUT_DIR}/clust_{tag}_mem', sizeratio=SIZE_RATIO):
         plt.bar(d['name'][:4], d['max_mem_mb'][:4])
         plt.bar(d['name'][4:], d['max_mem_mb'][4:])
         plt.ylabel('Max memory (MB)')
         plt.xticks(rotation=90)
-    with ctx(f'{OUT_DIR}/clust_{tag}_ari', sizeratio=0.75):
+    with ctx(f'{OUT_DIR}/clust_{tag}_ari', sizeratio=SIZE_RATIO):
[…]
