### 2025-12-26 d84c784
Extract blini algorithms to a library
